pyproject.toml
[tool.poetry]
name = "python_test_task"
version = "0.1.0"
description = ""
authors = ["Tony Benoy <me@tonybenoy.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8,<3.10"
attrs = "20.3.0"
numpy = "1.19.3"
opencv-python = "4.5.1.48"
tensorflow-hub = "0.11.0"
tensorflow = "2.7"
typer = "^0.6.1"
fastapi = "^0.85.0"
uvicorn = {extras = ["standard"], version = "^0.18.3"}
gunicorn = "^20.1.0"
SQLAlchemy = "^1.4.41"
psycopg2-binary = "^2.9.3"
rq = "^1.11.1"
alembic = "^1.8.1"
python-dotenv = "^0.21.0"

[tool.poetry.group.dev.dependencies]
mypy = "^0.981"
black = "^22.8.0"
flake8 = "^5.0.4"
pylint = "^2.15.3"
pre-commit = "^2.20.0"
isort = "^5.10.1"
pytest = "^7.1.3"

[tool.poetry.scripts]
webapp_test = "uvicorn webapp.webapp:app --reload"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
